backend/requirements.txt
flask
python-dotenv
bcrypt
pyjwt
werkzeug
gunicorn
